Include dead regions in graph traversal roots
This ensures that they will be included in graph printouts even if they
are unreachable in the CFG, as long as they are used. In general,
it makes more sense for the generic graph root finder to ensure that the
defs of any used values will be traversed, as that is typically the
desired property.