Stop constraining incoming edges in drawn graphs
The dot graphs output by the visualizer had `ordering=in` specified on
their nodes because this was how it originally guaranteed that input
edges would be displayed in the correct order. The visualizer was then
updated to use dedicated graphviz ports for node inputs (as this
resulted in clearer graphs), but the `ordering=in` constraint was never
removed.

This attribute resulted in many edges being drawn in much more
convoluted/complicated ways than necessary; removing it made quite a few
graphs look much better.