Fix dataflow verification of phi nodes
Previously, the verification did nothing as it forgot to walk down the
dominator tree. Fixing that bug revealed another, namely that the inputs
to phi nodes might not always be scheduled when the phi nodes are
checked, as the initial scheduling pass skips phi inputs. The phi checks
now ensure that inputs are scheduled before checking their schedule
positions.