Switch to block-based codegen scheduling
This fixes some problems with the previous scheduler design, including
providing a way for values to be scheduled before block headers, which
wasn't possible before; this should allow partially redundant
computations to be sunk more aggressively.

There was an  additional drive-by change here: the `BlockCfg` now
creates a new block whenever a region is encountered, to allow
scheduling to avoid having to deal with phi nodes in the middle of
blocks.