Switch back to linked lists for valgraph use lists
This makes the code somewhat simpler and easier to reason about, at the
expense of potentially less cache locality for long use chains and an
additional 4 bytes of overhead per use object. It does provide one less
level of indirection for single-use values, such as control values.

The transition to this new representation had the side-effect of
reversing all use lists relative to their previous order, which is what
caused the churn in most of the tests.