Remove single-input phis during isel lowering
This is a step in the transition away from outgoing block params for
multi-successor blocks. Because critical edges are expected to have
already been split during codegen, the only cases where a block with
multiple successors can have outgoing block params are if all successors
have a single predecessor, but the parameters are redundant in those
cases.

After this change, we no longer emit outgoing block parameters in those
cases, which should allow the complex multi-successor block param
support to be removed completely from the LIR.

The ultimate goal here is to give the register allocator a stricter set
of assumptions to work with; if no multi-successor block has outgoing
block params, the parallel copies for any block's block params can
always safely be placed in its predecessor.