Track live ranges of block params more accurately
Because copies for block params will always be inserted in the preceding
block, the block params themselves actually become live at the pre-copy
slot of every preceding jump instruction. Modelling things this
accurately isn't strictly necessary if the jumps can't produce output
values of their own (which we generally expect to be the case), but this
is safer until the rest of the allocator is in place.