Fix scheduling of block terminators
Previously, if data nodes were hoisted into a block with a control
terminator node (currently only possible with `brcond`), the hoisted
nodes would always be placed *after* the terminator, because they were
always added to the unscheduled node list after the terminator.

Now, block terminators are tracked separately during the pinning phase
and are manually attached to the end of the block schedule irrespective
of the rest of the sort.