Disallow all defs in block terminators in regalloc
We need this restriction even when there are no outgoing block params,
because live range splitting later can create situations where blocks
with no outgoing parameters still need copies to be inserted.

Technically, this is needed only when the block has one successor, as
that is when "edge copies" will need to be inserted before the jump
instead of in the relevant successor, but let's be more conservative for
now.