Implement basic hinting support in regalloc
The probe order of physical registers is now adjusted appropriately
whenever a register has been hinted for certain physical registers by
operand constraints. The new allocation order starts with hinted
registers (in order of decreasing weight), and falls back to the global
machine-defined allocation order once the hints have been exhausted.

There is still one glaring issue with things as implemented now: if a
virtual register hinted to a specific physical register by a use is live
after the instruction using it, it currently won't be allocated to that
physical register. The reason is that the physical register is currently
reserved early (during live range computation) for the relevant part of
the instruction, so the live range used by the hinted virtual register
will actually end up interfering with the previous reservation and force
a different register to be used.

An illustrative example of this happens with signed division on x86,
where the original LIR might look like this:

func @sdiv64 {
block0[%1:gpr($rdi), %2:gpr($rsi)]:
    %3:gpr($rdx)[late] = ConvertWord(Cqo) %1($rax)[early]
    %0:gpr($rax)[late], %4:gpr($rdx)[late] = Idiv(S64) %1($rax)[early], %3($rdx)[early], %2(any)[early]
    Ret %0($rax)[early]
}

There is a very strong hint tying `%1` to `$rax` here (it is used that
way both by the `cqo` and by the division itself), but the allocator
can't currently assign it to `$rax` because `$rax` is reserved during
the first part of the `cqo`, so it assigns `$rdi` instead, and we end up
with the equivalent of:

mov rax, rdi
cqo
mov rax, rdi
idiv rsi
ret

We'll need some way to deal with that situation in the future, but this
is a good start.