Move physical reservations out of allocation tree
Tracking register reservations separately means that ordinary
allocations can overlap them, which is what we need when assigning a
live range that is live-through an instruction to a physical register
copied out of it in the instruction. Now that reservations track the
live range they copy out of, the old `pre_instr_preg_copies` and
`post_instr_preg_copies` are no longer necessary and have been removed.