Extract x64 division emission
There are now functions responsible for emitting the necessary
signed/unsigned divide sequences that return both the quotient and
remainder. This should make it easier to emit remainders or combined
quotient/remainder operations in the future.

The fact that we now use `copy_vreg` to copy into the output register
caused some churn in the tests because of register renumbering, but
there is no functional change.