Fix splitting of ranges near end
Previously, attempting to split a range ending at a pre-copy point
directly before that pre-copy point would erroneously do nothing,
because the splitting logic thought it had hit the end of the range
instead of the point just before it. This could then cause infinite
loops as the same split was attempted to no avail.

We now correctly check whether the split point is inside the range or at
its boundary.