Store more information in `LiveRangeInstr`
The goal is for this structure to summarize just enough for
spills/reloads to be inserted correctly during fragment spilling. That
means we need to track:

* Whether the instruction even _wants_ the operand in a register. If
  not, it can be rewritten to a direct stack access later.

* Whether the operand is a use or a def, so we know whether to insert a
  spill or a reload.

* Where exactly within the instruction the value is used/defined, so we
  can create the right range for the value being spilled/reloaded when
  appropriate.

The implementation is slightly complicated by the fact that all operands
using a given vreg within an instruction are aggregated into a single
`LiveRangeInstr` instance so that spill weights make sense. When
aggregating uses that way, we also need to make sure that the recorded
use point and the `needs_reg` flag still make sense.

Another complication arises when dealing with fixed-register defs. Since
we don't have a dedicated "post-copy" slot, the definition point of the
output vreg is set to `Before` the next instruction, but using that
position directly within the `LiveRangeInstr` would cause it to be
attributed to the wrong instruction. To avoid that, we use a new enum
similar to `InstrSlot` that has an `After` point instead of a `Before`.
Once the allocator is more fleshed out there might be a cleaner way to
do this.