Stop storing scheduled nodes in reverse order
This was convenient given the current block scheduling implementation,
which "schedules" by taking an arbitrary topological sort of the block
using a postorder traversal. Almost every other scheduling algorithm,
however, (notably including list scheduling) produces the schedule in
forward order, so the whole exposure of the reversed schedule order was
a rather leaky abstraction. In preparation for a proper scheduling
heuristic, let's stop pretending the schedule is backward.

The test churn here was caused by stack slots being renumbered in the
wake of the schedule traversal changes.