Implement last-use-count scheduling heuristic
This heuristic, from the paper "Experimental evaluation of various
register-pressure-reduction heuristics", prioritizes nodes that have a
higher number of input values for which the node is the last use of
that input in the block. The idea is that the last-use-count roughly
approximates the number of live ranges that would end if the node were
placed (ignoring the fact that values may be live-out of the block), and
ending live ranges eagerly is good for reducing register pressure.

This heuristic alone still isn't great for register pressure, because
nodes that aren't last users can often "block" nodes that are and
prevent them from being ready. To counter that, ties for last-use-count
are broken using the existing critical path heuristic, which helps start
longer computation chains early and exposes more opportunities for
last-use-count.

This new heuristic yields optimal (RP-wise) scheduling on parts of the
TDN codegen bringup tests, which often call a bunch of functions and
accumulate their results together with a binary operation. A simplified
version of this pattern is now tested in the new
`accumulate_long_call_chain` test case.