Create "atomic" regalloc fragments
These are fragments that cover a single instruction and require a
register, so failing to allocate them is a hard error. To ensure that
non-atomic fragments are never allocated in favor of atomic ones, atomic
fragments get a special (effectively infinite) spill weight that is
higher than that of all non-atomic fragments.