Implement core regalloc spilling logic
When spilling, any operands that must reside in registers are moved into
their own atomic fragments for allocation later, and everything else is
left in the spilled ranges. In the future, we might not need to keep
around the spilled live ranges at all, which will allow at least some of
the code to be simplified.