Move `class` field out of live set fragments
The register class is actually a property of the entire live set, not of
an individual fragment, so there is no reason to duplicate it across all
fragments of a live set. We'll also need LiveSet -> class lookup in the
future, so let's keep that data in just one place.