Begin implementing regalloc parallel copy tracking
This is one of the last phases of allocation, after all live ranges have
already been assigned to registers and all operand assignments are
known: the various copies connecting different live ranges and blocks
need to be tracked so that they can be inserted into the final code.

Since some of these copies need to be performed "in parallel" with other
ones (i.e., permuting the value -> assignment mapping), they are tracked
in several "phases" per instruction, where the copies from each phase
are expected to be performed in parallel with one another. For final
codegen, these parallel copies will need to be broken up into a valid
sequence of pairwise copies with the same semantics.

Currently, only some of the necessary copies are inserted.