Rework regalloc assignment copy phases
The way the `Before` instruction slot was split into two phases (one for
reloads/fixed defs and one for everything else) turned out to be wrong,
because the core allocator assumes live ranges starting after a
fixed-def copy won't overlap with ones ending just before it. For
example, imagine the following (legal) scenario:

     %0     %1

     |
     |      !
   (rbx)  (rax: fixed-def)
     |      !
------------------------------ B:i7
     |      |
     |      |
   (rcx)  (rbx)
     |      |

Here, vreg %0 resides in $rbx until before i7, where it is moved to
$rcx, while %1 is copied from a fixed def into $rbx in that same slot -
both splitting and fixed-def handling use the `Before` slot for their
copies.

Previously, the dedicated `PrevInstrCleanup` handling would have caused
the following two parallel copies to be emitted:

    ($rbx) = ($rax)  PrevInstrCleanup
    ($rcx) = ($rbx)  CrossFragment

when in fact we need a single parallel copy to avoid clobbering the
value of %0 held in $rbx:

    ($rcx, $rbx) = ($rbx, $rax)

Fixed-def copies are thus now performed in the same phase as "standard"
cross-fragment copies. This shouldn't conflict with any other
`Before`-phase copies:

* Intra-block cross-fragment (split): non-interfering because fragments
  can't be split at their endpoints, and a fixed-def copy always starts
  a new fragment (it is a def).

* Block live-in copies: the previous instruction must be a terminator,
  but terminators can't have defs.

* Reloads: the only dangerous case is when the fixed def is copied into
  the spill slot being reloaded; discussed below.

Given that fixed-def copies actually can happen in parallel with other
cross-fragment copies, why did we need the extra phase in the first
place? The answer is reloads - if an instruction copies into a spill
slot and the instruction immediately after it reloads from that slot, we
need to make sure the spill happens before the reload, even though they
both use the successor's `Before` slot.

The existing code was an attempt to guarantee that (by putting spills in
the `PrevInstrCleanup` phase and reloads in the `CrossFragment` phase),
but it got some of the other phase assignments wrong.

The new code still uses the "middle" phase for reloads, but uses it
_only_ for reloads. This explains why the reload-from-fixed-def-spill
case above works correctly: if the spill and the reload reference the
same spill slot, the reload will always happen after the spill instead
of in parallel with it.

But wait, if we're still splitting the `Before` slot into two phases,
don't we still run the risk of causing issues similar to the one that
motivated this change? Because the `Reload` phase is used only for copies
from spill slots, the issue would manifest itself as a spill clobbering
an earlier value of a spill slot needed after the spill.

This can't happen for a rather subtle reason: consider two vreg ranges
sharing the same spill slot.

* If the ranges' vregs belong to the same live set, they won't ever
  expect different values, because a live set has at most one value at
  each program point.

* Otherwise, the ranges must have been assigned the same spill slot by
  spill slot coloring, meaning the spill hulls don't overlap. However, a
  `Reload`-phase copy never terminates a spill range because the spill
  range always extends at least past the use in the instruction
  (`PreCopy`, `Early` or `Late`). This means that no copy into the spill
  slot can occur in the `Before` instruction slot (it would interfere
  with the reloaded value), so the case described above can't happen.