Hook up free register probing to parallel copy resolution
The register allocator is now capable of finding a free register to
break cycles during parallel copy resolution/stack-to-stack copy
handling. The code still doesn't deal with the case of no free physical
registers at the copy point; in the future, we'll want to perform an
emergency spill to free something up.