Fix free register search during parallel copy resolution
Previously, any registers not allocated or reserved at the position of
the parallel copies being resolved were considered "free" for use as a
temporary. This logic failed to account for registers that had to be
live due to the order of the copies being emitted, even if they weren't
actually live at the relevant position.

In particular:

- Copy sources were considered available if their live ranges ended
  exactly at the copy point (i.e., they didn't have any later uses),
  even though they only become available after they have been copied out
  of in the lowered sequence.

- Copy destinations for block live-outs were considered available if
  they weren't live-in in the subsequent block, when in fact they become
  unavailable once copied into.

For now, take the rather heavy-handed approach of just reserving all
sources and destinations before resolving the copy. This is a strict
over-approximation of actual register liveness, but the cases where it
makes an observable difference should hopefully be rare.