Refactor machine emission interface a bit
It turns out that including frame info directly in the context passed to
the machine wasn't a great idea, since implementations often want to
hold on to bits of the frame info while emitting code bytes.

Instead, pass the info as a dedicated parameter to work around those
borrowing issues.