Correctly compensate for x64 stack adjustments
When pushing/popping parameters for calls, the stack pointer moves
relative to the base of the fixed frame, which means that any frame
accesses performed after setup/before teardown need to be adjusted as
well.

The current offset of the base of the frame from `rsp` is now tracked in
the `X64EmitState` and used when generating stack accesses.