Clarify that x64 ABS64 relocations have an addend in C API
This was always the intention, and is how `codegen_and_exec` handles
such relocations. We currently never actually emit a nonzero addend, but
it could be useful in the future if globals are implemented.

In the meantime let's make sure we're not painting ourselves into a
corner by requiring addends to be handled in all consumer code.