Fix handling of spill connectors during regalloc copy insertion
Spill/reload connector ranges are special in that they aren't actually
"canonical" ranges that have sole responsibility for carrying a value
across a particular part of the program. Instead, they always house an
in-register copy of the canonical value housed in a spill slot.

This means that for the purposes of intra-block copy handling,
connectors should never be considered as sources for copies into
subsequent ranges -- the true "previous" range is always the spill.

Failure to account for this caused copies out of long spill ranges
containing reloads not to be inserted, because the ranges after the
spills were never adjacent to the internal connectors.

Concretely, consider the following scenario (with all ranges depicted
belonging to a single vreg):

               S              R
    |---------------------|-------|
             |---|
               C

Here, S is assigned to a spill slot, C is a reload connector, and R is
assigned to a register. The expectation is that two copies will be
inserted: one from S to C (the reload for which C exists), and another
from S to R (an ordinary cross-range copy).

During reification, these ranges are sorted as

    S, C, R

with the expectation that C will be "folded into" S during the copy
insertion loop.

The old code, however, didn't treat C specially when tracking the "last"
range holding the value, so it ended up not inserting any copy into R
because C and R are not adjacent.

The fix is to skip connectors entirely when tracking which range holds a
value.