Expand regalloc assignment copy API
There is now more low-level access to the stored copies themselves, and
something reminiscent of the old `AssignmentCopyIter` has been
reintroduced as `AssignmentCopyTracker`.

The goal is to make more information available to consumers that need it
(such as the verifier and its users) without complicating the more
straightforward use cases.