Get rid of public `CodeBuffer::resolve_label` API
The reason this was exposed in the first place was mostly as a hack so
the x64 backend could use smaller branches when possible. That
optimization was removed in 30523e8, however, because it conflicted with
the new jump threading optimizations.

With the public `resolve_label` gone, the old `resolve_label_raw` could
just be called `resolve_label` itself. This necessitated some changes in
the tests, since the "raw" version doesn't sanitize its values and is
perfectly happy with returning an offset that that might later be moved
back due to branch pruning.