Restore `InstrSink` to its pre-87d6f28 definition
87d6f28 changed `InstrSink` to refer to a `SmallVec` instead of a `Vec`,
allowing it to be used for emission of reversed conditional branches.

The downside of the change was that it made common instruction emission
much slower, as they needed to be emitted into a `SmallVec` and then
copied into the `Vec`. With a bit of trickery, we can emit reversed
branches directly into the main `Vec` as well and then chop them off
when done.

This improved final emission performance by ~30% on the TDN monster
test.