Overhaul x64 constant selection/encoding
The different classes of `mov` (depending on bit size/signedness) are
now mostly resolved during isel instead of emission; this will allow
32-bit signed moves (other kinds aren't supported) directly to spill slots.

The one exception to the isel-level resolution is the s32 -> u32
transform for registers. The idea there is to let the operand be spilled
to memory by regalloc if it so desires, but still use the compact
encoding where possible for registers.