Start implementing IR graph canonicalization
This is the first proper "optimization' pass, and its goal is to
simplify the graph (constant folding, simple identities) and get it in a
form as canonical as possible for future passes.

For now, only a small subset of canonicalizations have been implemented,
and they don't yet hash-cons newly-constructed nodes. Still, this simple
subset should be powerful enough to do quite a few interesting things.