Add canonicalization support for `sfill`
This caused duplicate nodes to be created in the `ArrayJagged` test,
which threw off the `unordered` filetest matching (both attempted
matches matched the first node). Until we properly hash-cons them
together and figure out how to deal with intentional duplicates, make
sure there is a "barrier" between them to force the matches to be
distinct.