Upgrade most dependencies
This was just a mechanical `cargo upgrade -i` followed by the requisite
test updates.

Graphviz output tests broke because `FxHash` changed its internal
algorithm, causing the hash tables used for node/edge attributes to
return them in a different order.

The one dependency currently not upgraded was hashbrown, because it took
an API break and replaced `RawTable` with a new `HashTable` structure.