Refactor internal reducer state management
The context has now been split into the function body and "everything
else", as that makes it easier to create common utility functions
without running into borrow checker issues.

There was a small API break in that `ReducerBuilder` now has two
lifetime parameters, because it has a mutable reference to the common
state which has a mutable reference to the node cache.