Fix inverted regalloc split condition
Instead of splitting only fragments that cover use instructions, the
logic tried to split only ones that _don't_ have any uses. This can
never be the case with an unsplit fragment, so the net result was that
no splitting was ever performed.

It's interesting that things improved so drastically with no splitting
at all; restoring the intended split heuristic regressed many of those
improvements, so we'll need to tune things more carefully.