Add more regalloc and codegen tests
The TDN tests were all unoptimized, which resulted in large amounts of
redundant constants/computation. Things like that are great for
stressing the allocator, but we also want some more true-to-life test
cases to assess regalloc quality.

While we're at it, add some loop nest tests from TDN, as we don't have
any of those yet.