Add machine regalloc hook for rematerializability
After experimenting with dedicated tracking of this property in the LIR
(both per-vreg and per-instruction), this ended up feeling like the
cleanest way to do it (in part because it has no overhead at the LIR
level).

The asymmetry between operand constraints, which are explicitly encoded
in the LIR, and this out-of-band information provided later does feel a
little awkward, but it ended up being less awkward than the
alternatives.