Relax verifier constraints on remat destinations
As far as correctness is concerned, there is nothing wrong with
rematerializing into any kind of constrained operand, provided the
constraints are met. The fact that the allocator currently only
rematerializes `AnyReg` operands is an implementation detail and
shouldn't matter to the verifier.

Note that the verifier still expects all rematerializations to happen
into registers, even for `Any` destinations; supporting that would
require a bit more state, and the current checks are still sound.

We still disallow `Fixed` operands because those require some extra
checks as well.