Properly kill defs during remat spills
Now, whenever a rematerializable fragment containing the original
definition is spilled, the definition will be recorded as killed to
avoid pointless allocation/emission later.

Along the way, clean up `needs_reg` handling for rematerializable uses:
instead of setting the flag during the initial liveness walk, just set
it when creating spill connector ranges for the rematerialization, which
is where we actually need that property.