Mark more x64 instructions as rematerializable
We now rematerialize signed 32-bit constants and stack addresses as well,
which made for some dramatic codegen improvements. The most prominent of
these was the `bittest_run` family, which ended up losing all its
spills.

The more aggressive rematerialization did reveal several deficiencies
in the current implementation, however. The most egregious of these is
that we don't yet prune all dead definitions, leading to sequences like
this one:

    000042: 48 8d 54 24 18                  lea rdx, [rsp + 0x18]
    000047: 48 8d 54 24 1a                  lea rdx, [rsp + 0x1a]
    00004c: 48 8d 54 24 1c                  lea rdx, [rsp + 0x1c]
    000051: 48 8d 54 24 10                  lea rdx, [rsp + 0x10]
    000056: 48 8d 54 24 14                  lea rdx, [rsp + 0x14]
    00005b: 48 89 e2                        mov rdx, rsp
    00005e: 48 8d 54 24 08                  lea rdx, [rsp + 8]

We also don't yet trim rematerializable live ranges optimally when
splitting them, resulting in artifacts such as:

    00003a: ba 18 00 00 00                  mov edx, 0x18
    00003f: 48 89 d6                        mov rsi, rdx
    000042: 48 03 f0                        add rsi, rax

We'll need to deal with both in the future.