Treat rematerializable fragment hints as mandatory
When fragments can be completely rematerialized, it is better to let
them be split/spilled rather than be copied out of into another
register.

This resulted in generally reduced instruction count, reduced register
usage and reduced spills across all tests.

It did slightly regress code size in cases where `MovRI64` instructions
were rematerialized across calls. In the future, we might want to
distinguish between remats that are actually as cheap as copies, and
those that are just cheaper than loads.