Track multiple vregs per assignment in regalloc verifier
Previously, the regalloc verifier was stricter than necessary in its
handling of block exit ghost copies: whenever it encountered one, it
would forget the vreg currently attached to the indicated assignment and
track just the new one. In principle, that assignment could be used as
either of the vregs until overwritten, but the verifier accepted only
the destination vreg.

This strict behavior ended up working with assignments produced by the
allocator because it always placed block param copies in predecessors,
which meant the assignment referenced by the ghost copy was always a
freshly-overwritten one.

We'll soon want to sink those param copies into the successor where
possible, though, and at that point it will be possible for several
vregs to share a physical assignment.

Prepare for that now in the verifier by tracking lists of vregs per
assignment instead of just a single one. This ended up propagating all
the way to the public API because errors about unexpected values in
registers now report the full list of tracked vregs.