Reorganize conflict split point selection
There is now a single loop that searches for both the last instruction
below the prospective split point and the first instruction above it;
this will help trim the split ranges when the first half is completely
rematerializable.

In preparation for that, also keep track of whether the relevant part of
the fragment is rematerializable throughout the loop.