Clean up fragment atomicity checks
What we really care about when deciding whether a fragment is atomic is
how many instruction boundaries it crosses, so just check for that
instead of using the various heuristics we had before.

This let us get rid of `ProgramPoint::prev`, which is no longer used
anywhere, and cleaned up remaining redundant copies around
rematerializable fragments.