Change meaning of fragment `CHEAPLY_REMATTABLE` flag
It now means that all live ranges in the fragment can be rematerialized
cheaply, even if they don't come from the same source. The one place in
the code that still uses the flag didn't actually depend on that
property: it just wants to know whether the fragment is worth keeping
for the cost of a few copies, and the vregs comprising the fragment
don't affect that decision.