Get rid of x64 `MovRZ` instruction
Instead, treat it as a `MovRmS32` internally and make the final decision
about choosing a zeroing idiom during emission.

This had the (intended) side effect of making the constant 0
rematerializable, which improved codegen in several test cases.

Part of the trouble with making `MovRZ`-like instructions
rematerializable is that they might clobber flags at points where
they're live, so the backend now dynamically computes flag liveness
whenever it wants to emit the constant 0. This happens at most once per
block (and hopefully only in a very small percentage of blocks), so it
seems like a reasonable tradeoff.