Avoid collecting copy hints from conflicting fragments
Assignment won't be able to respect the intent of the hint in those
cases: either the current fragment will be assigned to a different
register, or the hint fragment will be evicted for the assignment,
defeating the point of the hint and creating extra work.

This currently regressed the canonicalized large_spills test a bit,
probably because some rematerializable fragment lost a hint. We'll have
to deal with that more carefully in the future.