Report conflicts from unsatisfied cheap remat hints
Now, if a cheaply-rematerializable fragment has copy hints, but none can
be satisfied to to conflicts with the hint fragments, the conflict will
be reported instead of the fragment just remaining unhinted.

This is similar to the existing logic that makes hints mandatory for
cheaply-rematerializable fragments: there isn't a reason to place a
rematerializable value in a register where it can never be used
directly.