Fix link-eval label updates in domtree computation
As documented in the code itself, the label of every node is supposed to
be a node `w` with minimal semidominator such that

    r -+-> w -*-> node,

where `r` is the root of the tree containing `node` in the forest.

However, when updating labels during path compression in the forest, we
ended up taking minimal _labels_ instead of labels with minimal
_semidominators_.

I'm not really sure how that ended up working so well across so many
cases in practice, but it finally ended up causing a node to be
scheduled illegally (i.e., in a block that did not actually dominate all
uses) in TDN's `testout1::Sub_Funclet_96` bringup test. The error was
then caught by the regalloc verifier, which noticed that the assigned
register could contain garbage at one of the later use sites.

Both the TDN test case and the problematic sub-graph extracted from it
have now been added as regression tests.