Avoid marking atomic fragments as "cheaply remattable"
Atomic fragments can never be rematerialized because they can't actually
be cut any smaller.

Various heuristics throughout the register allocator use the "cheaply
remattable" flag as an indicator that register hints should be mandatory
or that spilling should be preferred to copying. None of those things
actually hold for atomic fragments, though: if an atomic fragment fails
to allocate, the allocator just gives up and says "out of registers".

This finally came to a head in TDN's `Bool_And_Op::Sub_Funclet_148`
test, in which the rematerialized constant 0 failed to allocate because
it was copied along an edge into a register still in use at the jump
point. In theory, it could be better to try to rematerialize "on the
edge" instead, but we don't have a way to represent that within the
allocator.

So just change the definition of "cheaply remattable" to preclude atomic
fragments, which is what the code wanted to test for in practice, and
remove the now-redundant atomicity check in `should_spill_cheap_remat`.